pyproject.toml
[project]
name = "911der"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "groq>=0.33.0",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "pedantic>=2.3.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "supabase>=2.22.2",
    "uvicorn>=0.38.0",
    "twilio>=9.0.0",
    "langchain-groq>=0.0.1",
    "langchain-core>=0.1.26",
    "python-multipart>=0.0.20",
]
